api/requirements.txt
adal==1.2.1
amqp==2.4.2
asn1crypto==1.3.0
Babel==2.6.0
billiard==3.6.0.0
boto3==1.12.6
botocore==1.15.6
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
cssselect2==0.2.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.41.0
django-anymail==6.0
django-cors-headers==2.5.3
django-extra-fields==2.0.3
django-filter==2.1.0
django-phonenumber-field==2.3.1
django-rest-auth==0.9.5
django-simple-history==2.7.0
django-storages==1.9.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-extensions==0.4.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jmespath==0.9.5
kombu==4.5.0
Markdown==3.1
oauthlib==3.1.0
phonenumbers==8.10.9
Pillow==6.0.0
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.3
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sendgrid==6.4.3
six==1.12.0
sqlparse==0.3.0
starkbank-ecdsa==1.0.0
tinycss2==1.0.2
twilio==6.44.0
urllib3==1.24.1
vine==1.3.0
webencodings==0.5.1
whitenoise==5.0.1
